Enhance Metric Learning and GNN Inference Performance
- Added timing measurements for embedding, edge building, filtering, and GNN inference to optimize performance and identify bottlenecks.
- Improved tensor handling in `build_edges` to ensure inputs are float32 and device-compatible.
- Updated GNN filter functions to clone outputs and ensure correct data types, preventing unintended modifications.
- Implemented half-precision support for models when auto-casting is enabled, enhancing memory efficiency.
- Refactored edge feature calculations to ensure tensors are on the correct device and type.
- Added error handling for model precision conversions to maintain robustness.
- Cleaned up synchronization calls around CUDA operations for accurate timing.

+++ models/MetricLearning/4/interaction_gnn.py
@@ -288,16 +288,16 @@ class RecurrentInteractionGNN2(InteractionGNNParams, nn.Module):
 
         dst_index = torch.unsqueeze(dst, 0).tile(e_updated.shape[1], 1)
         edge_messages_from_src = torch.scatter_add(
-            torch.zeros(e_updated.shape[1], x.shape[0], device=dst_index.device),
+            torch.zeros(e_updated.shape[1], x.shape[0], device=dst_index.device, dtype=e_updated.dtype),
             1,
             dst_index,
             e_updated.T,
         ).T
         if not self.params["use_scatter_add"]:
             # use scatter_mean
-            e_ones = torch.ones(dst.shape[0], device=dst_index.device)
+            e_ones = torch.ones(dst.shape[0], device=dst_index.device, dtype=e_updated.dtype)
             count = torch.scatter_add(
-                torch.zeros(x.shape[0], device=dst_index.device),
+                torch.zeros(x.shape[0], device=dst_index.device, dtype=e_updated.dtype),
                 0,
                 dst,
                 e_ones,
@@ -307,16 +307,16 @@ class RecurrentInteractionGNN2(InteractionGNNParams, nn.Module):
 
         src_index = torch.unsqueeze(src, 0).tile(e_updated.shape[1], 1)
         edge_messages_from_dst = torch.scatter_add(
-            torch.zeros(e_updated.shape[1], x.shape[0], device=src_index.device),
+            torch.zeros(e_updated.shape[1], x.shape[0], device=src_index.device, dtype=e_updated.dtype),
             1,
             src_index,
             e_updated.T,
         ).T
         if not self.params["use_scatter_add"]:
             # use scatter_mean
-            e_ones = torch.ones(dst.shape[0], device=dst_index.device)
+            e_ones = torch.ones(dst.shape[0], device=dst_index.device, dtype=e_updated.dtype)
             count = torch.scatter_add(
-                torch.zeros(x.shape[0], device=src_index.device),
+                torch.zeros(x.shape[0], device=src_index.device, dtype=e_updated.dtype),
                 0,
                 src,
                 e_ones,
@@ -411,16 +411,16 @@ class ChainedInteractionGNN2(InteractionGNNParams, nn.Module):
 
             dst_index = torch.unsqueeze(dst, 0).tile(e_updated.shape[1], 1)
             edge_messages_from_src = torch.scatter_add(
-                torch.zeros(e_updated.shape[1], x.shape[0], device=dst_index.device),
+                torch.zeros(e_updated.shape[1], x.shape[0], device=dst_index.device, dtype=e_updated.dtype),
                 1,
                 dst_index,
                 e_updated.T,
             ).T
             if not self.params["use_scatter_add"]:
                 # use scatter_mean
-                e_ones = torch.ones(dst.shape[0], device=dst_index.device)
+                e_ones = torch.ones(dst.shape[0], device=dst_index.device, dtype=e_updated.dtype)
                 count = torch.scatter_add(
-                    torch.zeros(x.shape[0], device=dst_index.device),
+                    torch.zeros(x.shape[0], device=dst_index.device, dtype=e_updated.dtype),
                     0,
                     dst,
                     e_ones,
@@ -430,16 +430,16 @@ class ChainedInteractionGNN2(InteractionGNNParams, nn.Module):
 
             src_index = torch.unsqueeze(src, 0).tile(e_updated.shape[1], 1)
             edge_messages_from_dst = torch.scatter_add(
-                torch.zeros(e_updated.shape[1], x.shape[0], device=src_index.device),
+                torch.zeros(e_updated.shape[1], x.shape[0], device=src_index.device, dtype=e_updated.dtype),
                 1,
                 src_index,
                 e_updated.T,
             ).T
             if not self.params["use_scatter_add"]:
                 # use scatter_mean
-                e_ones = torch.ones(dst.shape[0], device=dst_index.device)
+                e_ones = torch.ones(dst.shape[0], device=dst_index.device, dtype=e_updated.dtype)
                 count = torch.scatter_add(
-                    torch.zeros(x.shape[0], device=src_index.device),
+                    torch.zeros(x.shape[0], device=src_index.device, dtype=e_updated.dtype),
                     0,
                     src,
                     e_ones,
